+++ src/test/groovy/ru/vyarus/dropwizard/guice/debug/yaml/BindingsReportTest.groovy
@@ -40,6 +40,15 @@ class BindingsReportTest extends Specification {
 
     Unique sub configuration objects bindings:
 
+        Configuration.admin
+            @Config AdminFactory = AdminFactory[healthChecks=HealthCheckConfiguration[minThreads=1, maxThreads=4, workQueueSize=1], tasks=TaskConfiguration[printStackTraceOnError=false]]
+
+        Configuration.admin.healthChecks
+            @Config HealthCheckConfiguration = HealthCheckConfiguration[minThreads=1, maxThreads=4, workQueueSize=1]
+
+        Configuration.admin.tasks
+            @Config TaskConfiguration = TaskConfiguration[printStackTraceOnError=false]
+
         Configuration.logging
             @Config LoggingFactory (with actual type DefaultLoggingFactory) = DefaultLoggingFactory{level=INFO, loggers={}, appenders=[io.dropwizard.logging.ConsoleAppenderFactory@1111111]}
 
@@ -62,6 +71,13 @@ class BindingsReportTest extends Specification {
     Configuration paths bindings:
 
         Configuration:
+            @Config("admin") AdminFactory = AdminFactory[healthChecks=HealthCheckConfiguration[minThreads=1, maxThreads=4, workQueueSize=1], tasks=TaskConfiguration[printStackTraceOnError=false]]
+            @Config("admin.healthChecks") HealthCheckConfiguration = HealthCheckConfiguration[minThreads=1, maxThreads=4, workQueueSize=1]
+            @Config("admin.healthChecks.maxThreads") Integer = 4
+            @Config("admin.healthChecks.minThreads") Integer = 1
+            @Config("admin.healthChecks.workQueueSize") Integer = 1
+            @Config("admin.tasks") TaskConfiguration = TaskConfiguration[printStackTraceOnError=false]
+            @Config("admin.tasks.printStackTraceOnError") Boolean = false
             @Config("logging") LoggingFactory (with actual type DefaultLoggingFactory) = DefaultLoggingFactory{level=INFO, loggers={}, appenders=[io.dropwizard.logging.ConsoleAppenderFactory@1111111]}
             @Config("logging.appenders") List<AppenderFactory<ILoggingEvent>> (with actual type ArrayList<AppenderFactory<ILoggingEvent>>) = [io.dropwizard.logging.ConsoleAppenderFactory@1111111]
             @Config("logging.level") String = "INFO"
@@ -121,6 +137,16 @@ class BindingsReportTest extends Specification {
     │
     ├── sub: ComplexGenericCase.Sub<String>
     │   └── smth: String = null
+    │
+    ├── admin: AdminFactory
+    │   │
+    │   ├── healthChecks: HealthCheckConfiguration
+    │   │   ├── maxThreads: Integer = 4
+    │   │   ├── minThreads: Integer = 1
+    │   │   └── workQueueSize: Integer = 1
+    │   │
+    │   └── tasks: TaskConfiguration
+    │       └── printStackTraceOnError: Boolean = false
     │
     ├── logging: DefaultLoggingFactory
     │   ├── appenders: ArrayList<AppenderFactory<ILoggingEvent>> = [io.dropwizard.logging.ConsoleAppenderFactory@1111111]
@@ -195,6 +221,15 @@ class BindingsReportTest extends Specification {
         ComplexGenericCase.sub
             @Config ComplexGenericCase.Sub<String> = null
 
+        Configuration.admin
+            @Config AdminFactory = AdminFactory[healthChecks=HealthCheckConfiguration[minThreads=1, maxThreads=4, workQueueSize=1], tasks=TaskConfiguration[printStackTraceOnError=false]]
+
+        Configuration.admin.healthChecks
+            @Config HealthCheckConfiguration = HealthCheckConfiguration[minThreads=1, maxThreads=4, workQueueSize=1]
+
+        Configuration.admin.tasks
+            @Config TaskConfiguration = TaskConfiguration[printStackTraceOnError=false]
+
         Configuration.logging
             @Config LoggingFactory (with actual type DefaultLoggingFactory) = DefaultLoggingFactory{level=INFO, loggers={}, appenders=[io.dropwizard.logging.ConsoleAppenderFactory@1111111]}
 
@@ -221,6 +256,13 @@ class BindingsReportTest extends Specification {
             @Config("sub.smth") String = null
 
         Configuration:
+            @Config("admin") AdminFactory = AdminFactory[healthChecks=HealthCheckConfiguration[minThreads=1, maxThreads=4, workQueueSize=1], tasks=TaskConfiguration[printStackTraceOnError=false]]
+            @Config("admin.healthChecks") HealthCheckConfiguration = HealthCheckConfiguration[minThreads=1, maxThreads=4, workQueueSize=1]
+            @Config("admin.healthChecks.maxThreads") Integer = 4
+            @Config("admin.healthChecks.minThreads") Integer = 1
+            @Config("admin.healthChecks.workQueueSize") Integer = 1
+            @Config("admin.tasks") TaskConfiguration = TaskConfiguration[printStackTraceOnError=false]
+            @Config("admin.tasks.printStackTraceOnError") Boolean = false
             @Config("logging") LoggingFactory (with actual type DefaultLoggingFactory) = DefaultLoggingFactory{level=INFO, loggers={}, appenders=[io.dropwizard.logging.ConsoleAppenderFactory@1111111]}
             @Config("logging.appenders") List<AppenderFactory<ILoggingEvent>> (with actual type ArrayList<AppenderFactory<ILoggingEvent>>) = [io.dropwizard.logging.ConsoleAppenderFactory@1111111]
             @Config("logging.level") String = "INFO"
@@ -330,6 +372,16 @@ class BindingsReportTest extends Specification {
                 .showConfigurationTreeOnly()) == """
 
     ComplexGenericCase (visible paths)
+    │
+    ├── admin: AdminFactory
+    │   │
+    │   ├── healthChecks: HealthCheckConfiguration
+    │   │   ├── maxThreads: Integer = 4
+    │   │   ├── minThreads: Integer = 1
+    │   │   └── workQueueSize: Integer = 1
+    │   │
+    │   └── tasks: TaskConfiguration
+    │       └── printStackTraceOnError: Boolean = false
     │
     ├── logging: DefaultLoggingFactory
     │   ├── appenders: ArrayList<AppenderFactory<ILoggingEvent>> = [io.dropwizard.logging.ConsoleAppenderFactory@1111111]