+++ src/test/groovy/ru/vyarus/dropwizard/guice/yaml/ConfigInspectorTest.groovy
@@ -11,8 +11,11 @@ import io.dropwizard.metrics.MetricsFactory
 import io.dropwizard.request.logging.RequestLogFactory
 import io.dropwizard.server.DefaultServerFactory
 import io.dropwizard.server.ServerFactory
+import io.dropwizard.servlets.tasks.TaskConfiguration
+import io.dropwizard.setup.AdminFactory
 import io.dropwizard.setup.Bootstrap
 import io.dropwizard.setup.Environment
+import io.dropwizard.setup.HealthCheckConfiguration
 import io.dropwizard.util.Duration
 import ru.vyarus.dropwizard.guice.GuiceBundle
 import ru.vyarus.dropwizard.guice.module.yaml.ConfigurationTree
@@ -41,7 +44,14 @@ class ConfigInspectorTest extends Specification {
         when: "check default config"
         def res = ConfigTreeBuilder.build(bootstrap, create(Configuration))
         then:
-        printConfig(res) == """[Configuration] logging (LoggingFactory as DefaultLoggingFactory) = DefaultLoggingFactory{level=INFO, loggers={}, appenders=[io.dropwizard.logging.ConsoleAppenderFactory@1111111]}
+        printConfig(res) == """[Configuration] admin (AdminFactory) = AdminFactory[healthChecks=HealthCheckConfiguration[minThreads=1, maxThreads=4, workQueueSize=1], tasks=TaskConfiguration[printStackTraceOnError=false]]
+[Configuration] admin.healthChecks (HealthCheckConfiguration) = HealthCheckConfiguration[minThreads=1, maxThreads=4, workQueueSize=1]
+[Configuration] admin.healthChecks.maxThreads (Integer) = 4
+[Configuration] admin.healthChecks.minThreads (Integer) = 1
+[Configuration] admin.healthChecks.workQueueSize (Integer) = 1
+[Configuration] admin.tasks (TaskConfiguration) = TaskConfiguration[printStackTraceOnError=false]
+[Configuration] admin.tasks.printStackTraceOnError (Boolean) = false
+[Configuration] logging (LoggingFactory as DefaultLoggingFactory) = DefaultLoggingFactory{level=INFO, loggers={}, appenders=[io.dropwizard.logging.ConsoleAppenderFactory@1111111]}
 [Configuration] logging.appenders (List<AppenderFactory<ILoggingEvent>> as ArrayList<AppenderFactory<ILoggingEvent>>) = [io.dropwizard.logging.ConsoleAppenderFactory@1111111]
 [Configuration] logging.level (String) = "INFO"
 [Configuration] logging.loggers (Map<String, JsonNode> as HashMap<String, JsonNode>) = {}
@@ -98,8 +108,8 @@ class ConfigInspectorTest extends Specification {
 [Configuration] server.umask (String) = null
 [Configuration] server.user (String) = null"""
         res.rootTypes == [Configuration]
-        res.uniqueTypePaths.size() == 6
-        res.paths.size() == 56
+        res.uniqueTypePaths.size() == 9
+        res.paths.size() == 63
         check(res, "server", DefaultServerFactory)
         check(res, "server.maxThreads", Integer, 1024)
         check(res, "server.idleThreadTimeout", Duration, Duration.minutes(1))
@@ -114,8 +124,8 @@ class ConfigInspectorTest extends Specification {
 [SimpleConfig] foo (String) = null
 [SimpleConfig] prim (Integer) = 0"""
         res.rootTypes == [SimpleConfig, Configuration]
-        res.uniqueTypePaths.size() == 6
-        res.paths.size() == 59
+        res.uniqueTypePaths.size() == 9
+        res.paths.size() == 66
         check(res, "foo", String)
         check(res, "bar", Boolean)
         check(res, "prim", Integer)
@@ -126,8 +136,8 @@ class ConfigInspectorTest extends Specification {
         then: "Object remain as declared type"
         printConfig(res) == "[ObjectPropertyConfig] sub (Object) = null"
         res.rootTypes == [ObjectPropertyConfig, Configuration]
-        res.uniqueTypePaths.size() == 6
-        res.paths.size() == 57
+        res.uniqueTypePaths.size() == 9
+        res.paths.size() == 64
         check(res, "sub", Object)
         elt.isObjectDeclaration()
         elt.declaredType == Object
@@ -157,10 +167,10 @@ class ConfigInspectorTest extends Specification {
 [ComplexConfig] sub.two (ComplexConfig.Parametrized<String>) = null
 [ComplexConfig] sub.two.list (List<String>) = null"""
         res.rootTypes == [ComplexConfig, ComplexConfig.Iface, Configuration]
-        res.uniqueTypePaths.size() == 7
+        res.uniqueTypePaths.size() == 10
         res.uniqueTypePaths.find { it.valueType == ComplexConfig.SubConfig } != null
         res.uniqueTypePaths.find { it.valueType == ComplexConfig.Parametrized } == null
-        res.paths.size() == 62
+        res.paths.size() == 69
         check(res, "sub", ComplexConfig.SubConfig)
         check(res, "sub.sub", String)
         check(res, "sub.two", ComplexConfig.Parametrized, null, String)
@@ -225,14 +235,15 @@ class ConfigInspectorTest extends Specification {
         def res = ConfigTreeBuilder.build(bootstrap, create(NotUniqueSubConfig))
         then:
         res.getUniqueTypePaths().collect { it.getDeclaredType() } as Set ==
-                [ServerPushFilterFactory, LoggingFactory, ServerFactory, GzipHandlerFactory, MetricsFactory, RequestLogFactory] as Set
+                [ServerPushFilterFactory, LoggingFactory, ServerFactory, GzipHandlerFactory, MetricsFactory, RequestLogFactory,
+                 TaskConfiguration, AdminFactory, HealthCheckConfiguration] as Set
         res.findByPath("sub1").getDeclaredType() == NotUniqueSubConfig.SubConfig
         res.findByPath("sub1.sub").getDeclaredType() == String
         res.findAllByType(NotUniqueSubConfig.SubConfig).collect {it.path} == ["sub1", "sub2", "sub3"]
         res.findAllFrom(NotUniqueSubConfig).collect {it.path} == ["sub1", "sub1.sub", "sub2", "sub2.sub", "sub3", "sub3.sub"]
-        res.findAllRootPaths().collect {it.path} == ["sub1", "sub2", "sub3", "logging", "metrics", "server"]
+        res.findAllRootPaths().collect {it.path} == ["sub1", "sub2", "sub3", "admin", "logging", "metrics", "server"]
         res.findAllRootPathsFrom(NotUniqueSubConfig).collect {it.path} == ["sub1", "sub2", "sub3"]
-        res.findAllRootPathsFrom(Configuration).collect {it.path} == ["logging", "metrics", "server"]
+        res.findAllRootPathsFrom(Configuration).collect {it.path} == ["admin", "logging", "metrics", "server"]
         res.valueByPath("not.exists") == null
         res.valueByPath("sub1") == null
         res.valueByType(NotUniqueSubConfig.SubConfig) == null
@@ -245,15 +256,16 @@ class ConfigInspectorTest extends Specification {
         res = ConfigTreeBuilder.build(bootstrap, config)
         then:
         res.getUniqueTypePaths().collect { it.getDeclaredType() } as Set ==
-                [ComplexConfig.SubConfig, ServerPushFilterFactory, LoggingFactory, ServerFactory, GzipHandlerFactory, MetricsFactory, RequestLogFactory] as Set
+                [ComplexConfig.SubConfig, ServerPushFilterFactory, LoggingFactory, ServerFactory, GzipHandlerFactory, MetricsFactory, RequestLogFactory,
+                 TaskConfiguration, AdminFactory, HealthCheckConfiguration] as Set
         res.findByPath("sub").getDeclaredType() == ComplexConfig.SubConfig
         res.findByPath("sub.sub").getDeclaredType() == String
         res.findAllByType(ComplexConfig.SubConfig).collect {it.path} == ["sub"]
         res.findAllByType(ComplexConfig.Parametrized).collect {it.path} == ["one", "sub.two"]
         res.findAllFrom(ComplexConfig).collect {it.path} == ["one", "one.list", "sub", "sub.sub", "sub.two", "sub.two.list"]
-        res.findAllRootPaths().collect {it.path} == ["one", "sub", "logging", "metrics", "server"]
+        res.findAllRootPaths().collect {it.path} == ["one", "sub", "admin", "logging", "metrics", "server"]
         res.findAllRootPathsFrom(ComplexConfig).collect {it.path} == ["one", "sub"]
-        res.findAllRootPathsFrom(Configuration).collect {it.path} == ["logging", "metrics", "server"]
+        res.findAllRootPathsFrom(Configuration).collect {it.path} == ["admin", "logging", "metrics", "server"]
         res.valueByPath("not.exists") == null
         res.valueByPath("sub") == null
         res.valueByType(ComplexConfig.SubConfig) == null